Conflicts:
	.gitignore
	configure.ac

+++ tester/tester.c
@@ -17,11 +17,17 @@
  */
 
 #include <stdio.h>
-#include "CUnit/TestRun.h"
-#include "CUnit/Automated.h"
 #include "linphonecore.h"
 #include "private.h"
 #include "liblinphone_tester.h"
+
+#if __clang__ || ((__GNUC__ == 4 && __GNUC_MINOR__ >= 6) || __GNUC__ > 4)
+#pragma GCC diagnostic push
+#endif
+#pragma GCC diagnostic ignored "-Wstrict-prototypes"
+
+#include "CUnit/TestRun.h"
+#include "CUnit/Automated.h"
 #if HAVE_CU_CURSES
 #include "CUnit/CUCurses.h"
 #endif
@@ -29,9 +35,21 @@
 #include <gtk/gtk.h>
 #endif
 
+#if __clang__ || ((__GNUC__ == 4 && __GNUC_MINOR__ >= 6) || __GNUC__ > 4)
+#pragma GCC diagnostic pop
+#endif
+
+#if _WIN32
+#define unlink _unlink
+#endif
+
 static bool_t liblinphone_tester_ipv6_enabled=FALSE;
 static int liblinphone_tester_keep_accounts_flag = 0;
-static int manager_count = 0;
+static int liblinphone_tester_keep_record_files = FALSE;
+static int liblinphone_tester_leak_detector_disabled = FALSE;
+int manager_count = 0;
+int leaked_objects_count = 0;
+const MSAudioDiffParams audio_cmp_params = {10,2000};
 
 const char* test_domain="sipopen.example.org";
 const char* auth_domain="sip.example.org";
@@ -40,7 +58,9 @@ const char* test_password="secret";
 const char* test_route="sip2.linphone.org";
 const char *userhostsfile = "tester_hosts";
 
- static void network_reachable(LinphoneCore *lc, bool_t reachable) {
+const char *liblinphone_tester_mire_id="Mire: Mire (synthetic moving picture)";
+
+static void network_reachable(LinphoneCore *lc, bool_t reachable) {
 	stats* counters;
 	ms_message("Network reachable [%s]",reachable?"TRUE":"FALSE");
 	counters = get_stats(lc);
@@ -90,6 +110,7 @@ static void auth_info_requested(LinphoneCore *lc, const char *realm, const char 
 
 void reset_counters( stats* counters) {
 	if (counters->last_received_chat_message) linphone_chat_message_unref(counters->last_received_chat_message);
+	if (counters->last_received_info_message) linphone_info_message_destroy(counters->last_received_info_message);
 	memset(counters,0,sizeof(stats));
 }
 
@@ -107,7 +128,9 @@ LinphoneCore* configure_lc_from(LinphoneCoreVTable* v_table, const char* path, c
 
 	if (file){
 		filepath = ms_strdup_printf("%s/%s", path, file);
-		BC_ASSERT_TRUE_FATAL(ortp_file_exist(filepath)==0);
+		if (ortp_file_exist(filepath) != 0) {
+			ms_fatal("Could not find file %s in path %s, did you configured resources directory correctly?", file, path);
+		}
 		config = lp_config_new_with_factory(NULL,filepath);
 	}
 
@@ -183,7 +206,7 @@ bool_t wait_for_list(MSList* lcs,int* counter,int value,int timeout_ms) {
 #endif
 			linphone_core_iterate((LinphoneCore*)(iterator->data));
 		}
-#ifdef WIN32
+#ifdef LINPHONE_WINDOWS_DESKTOP
 		{
 			MSG msg;
 			while (PeekMessage(&msg, NULL, 0, 0,1)){
@@ -224,9 +247,23 @@ LinphoneCoreManager *get_manager(LinphoneCore *lc){
 	return manager;
 }
 
+bool_t transport_supported(LinphoneTransportType transport) {
+	Sal *sal = sal_init();
+	bool_t supported = sal_transport_available(sal,(SalTransport)transport);
+	if (!supported) ms_message("TLS transport not supported, falling back to TCP if possible otherwise skipping test.");
+	sal_uninit(sal);
+	return supported;
+}
+
+
+static void display_status(LinphoneCore *lc, const char *status){
+	ms_message("display_status(): %s",status);
+}
+
 LinphoneCoreManager* linphone_core_manager_init(const char* rc_file) {
 	LinphoneCoreManager* mgr= ms_new0(LinphoneCoreManager,1);
 	char *rc_path = NULL;
+	char *hellopath = bc_tester_res("sounds/hello8000.wav");
 	mgr->number_of_cunit_error_at_creation = CU_get_number_of_failures();
 	mgr->v_table.registration_state_changed=registration_state_changed;
 	mgr->v_table.auth_info_requested=auth_info_requested;
@@ -246,10 +283,11 @@ LinphoneCoreManager* linphone_core_manager_init(const char* rc_file) {
 	mgr->v_table.network_reachable=network_reachable;
 	mgr->v_table.dtmf_received=dtmf_received;
 	mgr->v_table.call_stats_updated=call_stats_updated;
+	mgr->v_table.display_status=display_status;
 
 	reset_counters(&mgr->stat);
 	if (rc_file) rc_path = ms_strdup_printf("rcfiles/%s", rc_file);
-	mgr->lc=configure_lc_from(&mgr->v_table, bc_tester_read_dir_prefix, rc_path, mgr);
+	mgr->lc=configure_lc_from(&mgr->v_table, bc_tester_get_resource_dir_prefix(), rc_path, mgr);
 	linphone_core_manager_check_accounts(mgr);
 
 	manager_count++;
@@ -257,6 +295,8 @@ LinphoneCoreManager* linphone_core_manager_init(const char* rc_file) {
 #if TARGET_OS_IPHONE
 	linphone_core_set_ringer_device( mgr->lc, "AQ: Audio Queue Device");
 	linphone_core_set_ringback(mgr->lc, NULL);
+#elif __QNX__
+	linphone_core_set_playback_device(mgr->lc, "QSA: voice");
 #endif
 
 #ifdef VIDEO_ENABLED
@@ -276,17 +316,18 @@ LinphoneCoreManager* linphone_core_manager_init(const char* rc_file) {
 #endif
 
 
+	linphone_core_set_play_file(mgr->lc,hellopath); /*is also used when in pause*/
+	ms_free(hellopath);
+
 	if( manager_count >= 2){
-		char hellopath[512];
-		char *recordpath = ms_strdup_printf("%s/record_for_lc_%p.wav",bc_tester_writable_dir_prefix,mgr->lc);
+		char *recordpath = ms_strdup_printf("%s/record_for_lc_%p.wav",bc_tester_get_writable_dir_prefix(),mgr->lc);
 		ms_message("Manager for '%s' using files", rc_file ? rc_file : "--");
-		linphone_core_use_files(mgr->lc, TRUE);
-		snprintf(hellopath,sizeof(hellopath), "%s/sounds/hello8000.wav", bc_tester_read_dir_prefix);
-		linphone_core_set_play_file(mgr->lc,hellopath);
+		linphone_core_set_use_files(mgr->lc, TRUE);
 		linphone_core_set_record_file(mgr->lc,recordpath);
 		ms_free(recordpath);
 	}
-	linphone_core_set_user_certificates_path(mgr->lc,bc_tester_writable_dir_prefix);
+
+	linphone_core_set_user_certificates_path(mgr->lc,bc_tester_get_writable_dir_prefix());
 
 	if (rc_path) ms_free(rc_path);
 
@@ -316,7 +357,7 @@ void linphone_core_manager_start(LinphoneCoreManager *mgr, const char* rc_file, 
 
 	linphone_core_get_default_proxy(mgr->lc,&proxy);
 	if (proxy) {
-		mgr->identity = linphone_address_new(linphone_proxy_config_get_identity(proxy));
+		mgr->identity = linphone_address_clone(linphone_proxy_config_get_identity_address(proxy));
 		linphone_address_clean(mgr->identity);
 	}
 }
@@ -341,19 +382,30 @@ void linphone_core_manager_stop(LinphoneCoreManager *mgr){
 }
 
 void linphone_core_manager_destroy(LinphoneCoreManager* mgr) {
+	if (mgr->stat.last_received_chat_message) {
+		linphone_chat_message_unref(mgr->stat.last_received_chat_message);
+	}
+	if (mgr->stat.last_received_info_message) linphone_info_message_destroy(mgr->stat.last_received_info_message);
 	if (mgr->lc){
 		const char *record_file=linphone_core_get_record_file(mgr->lc);
-		if (record_file){
+		int unterminated_calls;
+
+		if (!liblinphone_tester_keep_record_files && record_file){
 			if ((CU_get_number_of_failures()-mgr->number_of_cunit_error_at_creation)>0) {
 				ms_message ("Test has failed, keeping recorded file [%s]",record_file);
 			} else {
 				unlink(record_file);
 			}
 		}
+		BC_ASSERT_EQUAL((unterminated_calls=ms_list_size(mgr->lc->calls)), 0, int, "%i");
+		if (unterminated_calls != 0) {
+			ms_error("There are still %d calls pending, please terminates them before invoking linphone_core_manager_destroy().", unterminated_calls);
+		}
 		linphone_core_destroy(mgr->lc);
 	}
-	if (mgr->identity) linphone_address_destroy(mgr->identity);
-	if (mgr->stat.last_received_chat_message) linphone_chat_message_unref(mgr->stat.last_received_chat_message);
+	if (mgr->identity) {
+		linphone_address_destroy(mgr->identity);
+	}
 	manager_count--;
 
 	ms_free(mgr);
@@ -367,7 +419,7 @@ int liblinphone_tester_ipv6_available(void){
 		socklen_t slen=sizeof(ss);
 		char localip[128];
 		int port=0;
-		belle_sip_get_src_addr_for(ai->ai_addr,ai->ai_addrlen,(struct sockaddr*) &ss,&slen,4444);
+		belle_sip_get_src_addr_for(ai->ai_addr,(socklen_t)ai->ai_addrlen,(struct sockaddr*) &ss,&slen,4444);
 		src.ai_addr=(struct sockaddr*) &ss;
 		src.ai_addrlen=slen;
 		belle_sip_addrinfo_to_ip(&src,localip, sizeof(localip),&port);
@@ -379,6 +431,14 @@ int liblinphone_tester_ipv6_available(void){
 
 void liblinphone_tester_keep_accounts( int keep ){
 	liblinphone_tester_keep_accounts_flag = keep;
+}
+
+void liblinphone_tester_keep_recorded_files(int keep){
+	liblinphone_tester_keep_record_files = keep;
+}
+
+void liblinphone_tester_disable_leak_detector(int disabled){
+	liblinphone_tester_leak_detector_disabled = disabled;
 }
 
 void liblinphone_tester_clear_accounts(void){
@@ -402,11 +462,210 @@ void liblinphone_tester_add_suites() {
 	bc_tester_add_suite(&remote_provisioning_test_suite);
 	bc_tester_add_suite(&quality_reporting_test_suite);
 	bc_tester_add_suite(&log_collection_test_suite);
-	bc_tester_add_suite(&transport_test_suite);
+	bc_tester_add_suite(&tunnel_test_suite);
 	bc_tester_add_suite(&player_test_suite);
 	bc_tester_add_suite(&dtmf_test_suite);
 #if defined(VIDEO_ENABLED) && defined(HAVE_GTK)
 	bc_tester_add_suite(&video_test_suite);
 #endif
 	bc_tester_add_suite(&multicast_call_test_suite);
+	bc_tester_add_suite(&proxy_config_test_suite);
+#if HAVE_SIPP
+	bc_tester_add_suite(&complex_sip_call_test_suite);
+#endif
 }
+
+static int linphone_core_manager_get_max_audio_bw_base(const int array[],int array_size) {
+	int i,result=0;
+	for (i=0; i<array_size; i++) {
+		result = MAX(result,array[i]);
+	}
+	return result;
+}
+
+static int linphone_core_manager_get_mean_audio_bw_base(const int array[],int array_size) {
+	int i,result=0;
+	for (i=0; i<array_size; i++) {
+		result += array[i];
+	}
+	return result/array_size;
+}
+
+int linphone_core_manager_get_max_audio_down_bw(const LinphoneCoreManager *mgr) {
+	return linphone_core_manager_get_max_audio_bw_base(mgr->stat.audio_download_bandwidth
+			, sizeof(mgr->stat.audio_download_bandwidth)/sizeof(int));
+}
+int linphone_core_manager_get_max_audio_up_bw(const LinphoneCoreManager *mgr) {
+	return linphone_core_manager_get_max_audio_bw_base(mgr->stat.audio_upload_bandwidth
+			, sizeof(mgr->stat.audio_upload_bandwidth)/sizeof(int));
+}
+
+int linphone_core_manager_get_mean_audio_down_bw(const LinphoneCoreManager *mgr) {
+	return linphone_core_manager_get_mean_audio_bw_base(mgr->stat.audio_download_bandwidth
+			, sizeof(mgr->stat.audio_download_bandwidth)/sizeof(int));
+}
+int linphone_core_manager_get_mean_audio_up_bw(const LinphoneCoreManager *mgr) {
+	return linphone_core_manager_get_mean_audio_bw_base(mgr->stat.audio_upload_bandwidth
+			, sizeof(mgr->stat.audio_upload_bandwidth)/sizeof(int));
+}
+
+void liblinphone_tester_before_each(void) {
+	if (!liblinphone_tester_leak_detector_disabled){
+		belle_sip_object_enable_leak_detector(TRUE);
+		leaked_objects_count = belle_sip_object_get_object_count();
+	}
+}
+
+static char* all_leaks_buffer = NULL;
+
+void liblinphone_tester_after_each(void) {
+	if (!liblinphone_tester_leak_detector_disabled){
+		int leaked_objects = belle_sip_object_get_object_count() - leaked_objects_count;
+		if (leaked_objects > 0) {
+			char* format = ms_strdup_printf("%d object%s leaked in suite [%s] test [%s], please fix that!",
+											leaked_objects, leaked_objects>1?"s were":"was",
+											bc_tester_current_suite_name(), bc_tester_current_test_name());
+			belle_sip_object_dump_active_objects();
+			belle_sip_object_flush_active_objects();
+			bc_tester_printf(bc_printf_verbosity_info, format);
+			ms_error("%s", format);
+
+			all_leaks_buffer = ms_strcat_printf(all_leaks_buffer, "\n%s", format);
+		}
+	}
+
+	if (manager_count != 0) {
+		ms_fatal("%d Linphone core managers are still alive!", manager_count);
+	}
+}
+
+void liblinphone_tester_uninit(void) {
+	// show all leaks that happened during the test
+	if (all_leaks_buffer) {
+		bc_tester_printf(bc_printf_verbosity_info, all_leaks_buffer);
+		ms_free(all_leaks_buffer);
+		all_leaks_buffer = NULL;
+	}
+	bc_tester_uninit();
+}
+
+static void check_ice_from_rtp(LinphoneCall *c1, LinphoneCall *c2, LinphoneStreamType stream_type) {
+	MediaStream *ms;
+	switch (stream_type) {
+	case LinphoneStreamTypeAudio:
+		ms=&c1->audiostream->ms;
+		break;
+	case LinphoneStreamTypeVideo:
+		ms=&c1->videostream->ms;
+		break;
+	case LinphoneStreamTypeText:
+		ms=&c1->textstream->ms;
+		break;
+	default:
+		ms_error("Unknown stream type [%s]",  linphone_stream_type_to_string(stream_type));
+		BC_ASSERT_FALSE(stream_type >= LinphoneStreamTypeUnknown);
+		return;
+	}
+	
+	
+	if (linphone_call_get_audio_stats(c1)->ice_state == LinphoneIceStateHostConnection && media_stream_started(ms)) {
+		char ip[16];
+		char port[8];
+		getnameinfo((const struct sockaddr *)&c1->audiostream->ms.sessions.rtp_session->rtp.gs.rem_addr
+					, c1->audiostream->ms.sessions.rtp_session->rtp.gs.rem_addrlen
+					, ip
+					, sizeof(ip)
+					, port
+					, sizeof(port)
+					, NI_NUMERICHOST|NI_NUMERICSERV);
+		BC_ASSERT_STRING_EQUAL(ip, c2->media_localip);
+	}
+}
+bool_t check_ice(LinphoneCoreManager* caller, LinphoneCoreManager* callee, LinphoneIceState state) {
+	LinphoneCall *c1,*c2;
+	bool_t audio_success=FALSE;
+	bool_t video_success=FALSE;
+	bool_t text_success=FALSE;
+	bool_t video_enabled, realtime_text_enabled;
+	MSTimeSpec ts;
+	
+	c1=linphone_core_get_current_call(caller->lc);
+	c2=linphone_core_get_current_call(callee->lc);
+	
+	BC_ASSERT_PTR_NOT_NULL(c1);
+	BC_ASSERT_PTR_NOT_NULL(c2);
+	if (!c1 || !c2) return FALSE;
+	linphone_call_ref(c1);
+	linphone_call_ref(c2);
+	
+	BC_ASSERT_EQUAL(linphone_call_params_video_enabled(linphone_call_get_current_params(c1)),linphone_call_params_video_enabled(linphone_call_get_current_params(c2)), int, "%d");
+	BC_ASSERT_EQUAL(linphone_call_params_realtime_text_enabled(linphone_call_get_current_params(c1)),linphone_call_params_realtime_text_enabled(linphone_call_get_current_params(c2)), int, "%d");
+	video_enabled=linphone_call_params_video_enabled(linphone_call_get_current_params(c1));
+	realtime_text_enabled=linphone_call_params_realtime_text_enabled(linphone_call_get_current_params(c1));
+	liblinphone_tester_clock_start(&ts);
+	do{
+		if ((c1 != NULL) && (c2 != NULL)) {
+			if (linphone_call_get_audio_stats(c1)->ice_state==state &&
+				linphone_call_get_audio_stats(c2)->ice_state==state ){
+				audio_success=TRUE;
+				check_ice_from_rtp(c1,c2,LinphoneStreamTypeAudio);
+				check_ice_from_rtp(c2,c1,LinphoneStreamTypeAudio);
+				break;
+			}
+			linphone_core_iterate(caller->lc);
+			linphone_core_iterate(callee->lc);
+		}
+		ms_usleep(20000);
+	}while(!liblinphone_tester_clock_elapsed(&ts,10000));
+	
+	if (video_enabled){
+		liblinphone_tester_clock_start(&ts);
+		do{
+			if ((c1 != NULL) && (c2 != NULL)) {
+				if (linphone_call_get_video_stats(c1)->ice_state==state &&
+					linphone_call_get_video_stats(c2)->ice_state==state ){
+					video_success=TRUE;
+					check_ice_from_rtp(c1,c2,LinphoneStreamTypeVideo);
+					check_ice_from_rtp(c2,c1,LinphoneStreamTypeVideo);
+					break;
+				}
+				linphone_core_iterate(caller->lc);
+				linphone_core_iterate(callee->lc);
+			}
+			ms_usleep(20000);
+		}while(!liblinphone_tester_clock_elapsed(&ts,10000));
+	}
+	
+	if (realtime_text_enabled){
+		liblinphone_tester_clock_start(&ts);
+		do{
+			if ((c1 != NULL) && (c2 != NULL)) {
+				if (linphone_call_get_text_stats(c1)->ice_state==state &&
+					linphone_call_get_text_stats(c2)->ice_state==state ){
+					text_success=TRUE;
+					check_ice_from_rtp(c1,c2,LinphoneStreamTypeText);
+					check_ice_from_rtp(c2,c1,LinphoneStreamTypeText);
+					break;
+				}
+				linphone_core_iterate(caller->lc);
+				linphone_core_iterate(callee->lc);
+			}
+			ms_usleep(20000);
+		}while(!liblinphone_tester_clock_elapsed(&ts,10000));
+	}
+	
+	/*make sure encryption mode are preserved*/
+	if (c1) {
+		const LinphoneCallParams* call_param = linphone_call_get_current_params(c1);
+		BC_ASSERT_EQUAL(linphone_call_params_get_media_encryption(call_param),linphone_core_get_media_encryption(caller->lc), int, "%d");
+	}
+	if (c2) {
+		const LinphoneCallParams* call_param = linphone_call_get_current_params(c2);
+		BC_ASSERT_EQUAL(linphone_call_params_get_media_encryption(call_param),linphone_core_get_media_encryption(callee->lc), int, "%d");
+	}
+	linphone_call_unref(c1);
+	linphone_call_unref(c2);
+	return video_enabled ? (realtime_text_enabled ? text_success && audio_success && video_success : audio_success && video_success) : realtime_text_enabled ? text_success && audio_success : audio_success;
+}
+
+
